One step kernel dwt forward (columns).

+++ src/matrix/dwt/tests/t_dwt2.cpp
@@ -11,7 +11,7 @@
 /**********************
  ** type definitions **
  **********************/
-typedef double elem_type;
+typedef float elem_type;
 
 
 using namespace codeare::matrix::io;
